move integration base type into common module
