+++ common/src/main/java/net/draycia/carbon/common/integration/Integration.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package net.draycia.carbon.paper.integration;
+package net.draycia.carbon.common.integration;
 
 public interface Integration {
